Code to push public weather data to prometheus, added some variables to simplify pushing data to proemetheus, bugfixes

+++ src/logger.py
@@ -3,26 +3,31 @@ import os
 import shutil
 
 
-def rotate_log_file(log_file_path, backup_log_file_path):
-    if os.path.exists(log_file_path):
+def rotate_log_file(log_file_path, backup_log_file_path, log_folder_path):
+    if os.path.exists(log_folder_path):
         if os.path.exists(backup_log_file_path):
             with open(backup_log_file_path, 'w'):
                 pass
         else:
+            print("app_old.log does not exist. Creating now...")
             with open(backup_log_file_path, 'w'):
                 pass
         
-        shutil.copy(log_file_path, backup_log_file_path)
+        if os.path.exists(log_file_path):
+            shutil.copy(log_file_path, backup_log_file_path)
 
         with open(log_file_path, 'w'):
             pass
+    else:
+        print("Error in accessing log folder, no logs will be saved.")
 
 
 root_folder = os.path.abspath(os.path.join(os.path.dirname(__file__), '..'))
+log_folder_path = os.path.join(root_folder, 'log')
 log_file_path = os.path.join(root_folder, 'log', 'app.log')
 old_log_file_path = os.path.join(root_folder, 'log', 'app_old.log')
 
-rotate_log_file(log_file_path, old_log_file_path)
+rotate_log_file(log_file_path, old_log_file_path, log_folder_path)
 
 logger = logging.getLogger('datasience')
 logger.setLevel(logging.DEBUG)
@@ -31,7 +36,7 @@ console_handler = logging.StreamHandler()
 file_handler = logging.FileHandler(log_file_path)
 
 console_handler.setLevel(logging.WARNING)
-file_handler.setLevel(logging.DEBUG)
+file_handler.setLevel(logging.INFO)
 
 formatter = logging.Formatter('%(asctime)s - %(levelname)s - %(message)s')
